api/forecast/requirements.txt
pandas
statsmodels
scikit-learn
python-dotenv
supabase-py
pmdarima
